moved org-select feature to viewer-button
