+++ client/src/app/components/layout/viewer-button/viewer-button.component.html
@@ -3,6 +3,7 @@
     nzDirection="horizontal"
     nzSize="middle"
     *ngrxLet="unreadCount$ as unreadCount">
+    <!-- ADD ENTITY DROPDOWN -->
     <button
       class="add-btn"
       nz-button
@@ -19,6 +20,8 @@
         nzType="caret-down"
         nzTheme="outline"></i>
     </button>
+
+    <!-- NOTIFICATION BUTTON -->
     <nz-badge
       [nzCount]="unreadCount"
       [nzOverflowCount]="999"
@@ -41,6 +44,8 @@
           nzTheme="outline"></i>
       </button>
     </nz-badge>
+
+    <!-- VIEWER BUTTON -->
     <button
       class="viewer-btn"
       nz-button
@@ -76,50 +81,90 @@
     </button>
   </nz-space>
 
+  <!-- VIEWER BUTTON DROPDOWN MENU -->
   <nz-dropdown-menu #userMenu="nzDropdownMenu">
     <ul nz-menu>
+      <!-- INVALID COI STATEMENT GROUP -->
       <li
-        nz-menu-item
-        *ngIf="viewer.invalidCoi">
-        <button
-          (click)="this.coiUpdateModalVisible = true"
-          nz-button
-          nzType="primary"
-          nzShape="round"
-          nzSize="small"
-          nzDanger
-          nzBlock>
-          <i
-            nz-icon
-            nzType="exclamation-circle"></i>
-          Please update COI statement
-        </button>
+        nz-menu-group
+        [nzTitle]="coiGroupTitle">
+        <ng-template #coiGroupTitle>
+          <span
+            nz-typography
+            nzType="danger">
+            <i
+              nz-icon
+              nzType="exclamation-circle"></i>
+            Conflict of Interest Invalid
+          </span>
+        </ng-template>
+        <ul *ngIf="viewer.invalidCoi">
+          <li nz-menu-item>
+            <button
+              *ngIf="viewer.invalidCoi"
+              (click)="coiUpdateModalVisible = true"
+              nz-button
+              nzType="primary"
+              nzShape="round"
+              nzSize="small"
+              nzDanger
+              nzBlock>
+              Please Update COI statement
+            </button>
+          </li>
+        </ul>
       </li>
       <li
-        nz-menu-divider
-        *ngIf="viewer.invalidCoi"></li>
-      <li
-        [routerLink]="['/users', viewer.id]"
-        nz-menu-item>
-        Your Profile
+        nz-menu-group
+        nzTitle="Curation">
+        <ul>
+          <li
+            *ngIf="viewer.organizations.length > 1"
+            nz-submenu
+            nzTitle="Change Organization">
+            <ul>
+              <li
+                *ngFor="let org of viewer.organizations"
+                (click)="menuSelection$.next(org.id)"
+                [ngClass]="{ selected: viewer.mostRecentOrg?.id === org.id }"
+                nz-menu-item>
+                <nz-avatar
+                  [nzSrc]="org?.profileImagePath"
+                  [nzSize]="14"
+                  [nzShape]="'square'"></nz-avatar>
+                <span class="org-name">{{ org.name }}</span>
+              </li>
+            </ul>
+          </li>
+          <li
+            nz-menu-item
+            nzMatchRouter
+            [routerLink]="['/users', viewer.id]">
+            Edit Profile
+          </li>
+          <li
+            nz-menu-item
+            (click)="signOut()">
+            Sign Out
+          </li>
+        </ul>
       </li>
       <li
-        nz-menu-item
-        href="/admin"
+        nz-menu-group
+        nzTitle="Administration"
         *ngIf="viewer.isAdmin">
-        <a href="/admin"> Admin Console </a>
-      </li>
-      <li
-        nz-menu-item
-        href="/jobs"
-        *ngIf="viewer.isAdmin">
-        <a href="/jobs"> Background Workers </a>
-      </li>
-      <li nz-menu-divider></li>
-      <li
-        nz-menu-item
-        (click)="signOut()">
-        Sign Out
+        <ul>
+          <li
+            nz-menu-item
+            href="/admin">
+            <a href="/admin"> Admin Console </a>
+          </li>
+          <li
+            nz-menu-item
+            href="/jobs">
+            <a href="/jobs"> Background Workers </a>
+          </li>
+        </ul>
       </li>
     </ul>
   </nz-dropdown-menu>
