+++ client/src/app/components/layout/viewer-button/viewer-button.component.ts
@@ -1,30 +1,33 @@
-import { Component, Input } from '@angular/core'
-import { Observable } from 'rxjs'
+import { Component, Input, OnInit } from '@angular/core'
 import { Viewer, ViewerService } from '@app/core/services/viewer/viewer.service'
-import { Maybe, ViewerNotificationCountGQL } from '@app/generated/civic.apollo'
-import { startWith, map } from 'rxjs/operators'
-import { NzModalRef, NzModalService } from 'ng-zorro-antd/modal'
+import { ViewerNotificationCountGQL } from '@app/generated/civic.apollo'
+import { Apollo, gql } from 'apollo-angular'
 import { environment } from 'environments/environment'
+import { Observable, Subject } from 'rxjs'
+import { map, startWith, withLatestFrom } from 'rxjs/operators'
 
 @Component({
   selector: 'cvc-viewer-button',
   templateUrl: './viewer-button.component.html',
   styleUrls: ['./viewer-button.component.less'],
 })
-export class CvcViewerButtonComponent {
+export class CvcViewerButtonComponent implements OnInit {
   @Input() cvcCollapsed: boolean = false
 
   viewer$: Observable<Viewer>
   unreadCount$: Observable<number>
 
+  menuSelection$: Subject<number>
   coiUpdateModalVisible: boolean = false
   addVariantModalVisible: boolean = false
 
   constructor(
     private queryService: ViewerService,
-    private unreadCountGql: ViewerNotificationCountGQL
+    private unreadCountGql: ViewerNotificationCountGQL,
+    private apollo: Apollo
   ) {
     this.viewer$ = this.queryService.viewer$
+    this.menuSelection$ = new Subject()
     if (environment.production) {
       this.unreadCount$ = this.unreadCountGql
         .watch(undefined, { pollInterval: 5000 })
@@ -42,6 +45,24 @@ export class CvcViewerButtonComponent {
     }
   }
 
+  ngOnInit(): void {
+    this.menuSelection$
+      .pipe(withLatestFrom(this.viewer$))
+      .subscribe(([mroId, viewer]: [number, Viewer]) => {
+        const fragment = {
+          id: `User:${viewer.id}`,
+          fragment: gql`
+            fragment UserMostRecentOrgId on User {
+              mostRecentOrganizationId
+            }
+          `,
+          data: {
+            mostRecentOrganizationId: mroId,
+          },
+        }
+        this.apollo.client.writeFragment(fragment)
+      })
+  }
   signOut(): void {
     this.queryService.signOut()
   }
