+++ client/src/app/forms2/types/org-submit-button/org-submit-button.type.html
@@ -4,67 +4,10 @@
   [formControl]="formControl"
   [formlyAttributes]="field" />
 
-<!-- base submit button -->
-<ng-template #submitButton>
-  <button
-    cvcOrgSubmitButton
-    type="submit"
-    nz-button
-    [disabled]="!form.valid"
-    [nzDanger]=""
-    nzType="primary">
-    {{ props.submitLabel }}
-  </button>
-</ng-template>
-
-<ng-container *ngIf="organizations$ | ngrxPush as organizations">
-  <!-- if 0 or 1 organization, display button template w/o btn-group wrapper -->
-  <ng-container *ngIf="organizations.length <= 1">
-    <ng-container *ngTemplateOutlet="submitButton"></ng-container>
-  </ng-container>
-
-  <!-- if multiple organizations, wrap in btn-group w/ selector btn -->
-  <ng-container *ngIf="organizations.length > 1">
-    <nz-button-group>
-      <ng-container *ngTemplateOutlet="submitButton"></ng-container>
-
-      <button
-        type="button"
-        nz-button
-        [ngClass]="(buttonClass$ | ngrxPush)!"
-        [hidden]="isHidden$ | ngrxPush"
-        [disabled]="isDisabled$ | ngrxPush"
-        nz-dropdown
-        nzTrigger="click"
-        [nzDropdownMenu]="orgMenu">
-        <span>for</span>
-        <ng-container *ngrxLet="mostRecentOrg$ as mrOrg">
-          <nz-avatar
-            [nzSrc]="mrOrg ? mrOrg.profileImagePath : organizations[0].profileImagePath"
-            [nzSize]="16"
-            [nzShape]="'square'">
-          </nz-avatar>
-        </ng-container>
-        <i
-          nz-icon
-          nzType="down"></i>
-      </button>
-    </nz-button-group>
-  </ng-container>
-
-  <!-- organization dropdown menu template -->
-  <nz-dropdown-menu #orgMenu="nzDropdownMenu">
-    <ul nz-menu>
-      <li
-        *ngFor="let org of organizations"
-        (click)="menuSelection$.next(org.id)"
-        nz-menu-item>
-        <nz-avatar
-          [nzSrc]="org?.profileImagePath"
-          [nzSize]="10"
-          [nzShape]="'square'"></nz-avatar>
-        {{ org.name }}
-      </li>
-    </ul>
-  </nz-dropdown-menu>
-</ng-container>
+<button
+  type="submit"
+  nz-button
+  [disabled]="!form.valid"
+  nzType="primary">
+  {{ props.submitLabel }}
+</button>
