+++ client/src/app/forms2/types/org-submit-button/org-submit-button.type.less
@@ -1,25 +1,3 @@
 :host {
   display: inline-block;
-  // fixes the right border on the first button
-  ::ng-deep .ant-btn-dangerous.ant-btn-primary:first-child:not(:last-child) {
-    border-right-color: #ff4d4f;
-  }
-  ::ng-deep .ant-btn-dangerous.ant-btn-primary:hover,
-  ::ng-deep .ant-btn-dangerous.ant-btn-primary:focus {
-    border-right-color: #fd7978;
-  }
-}
-
-.org-dropdown-btn {
-  nz-avatar {
-    margin: 0 6px;
-    margin-top: -2px;
-    border: 1px solid rgba(255,255,255,.3);
-    background-color: rgba(255,255,255,.9);
-  }
-  // fixes the left border on the last button
-  &.ant-btn-dangerous:last-child:not(:first-child),
-  .ant-btn-dangerous + .ant-btn-dangerous {
-    border-left-color: #fd7978;
-  }
 }
