+++ client/src/app/forms2/types/org-submit-button/org-submit-button.type.module.ts
@@ -1,18 +1,22 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { ReactiveFormsModule } from '@angular/forms';
-import { LetDirective, PushPipe } from '@ngrx/component';
-import { FormlyModule } from '@ngx-formly/core';
-import { ConfigOption } from '@ngx-formly/core/lib/models';
-import { NzAvatarModule } from 'ng-zorro-antd/avatar';
-import { NzButtonModule } from 'ng-zorro-antd/button';
-import { NzDropDownModule } from 'ng-zorro-antd/dropdown';
-import { NzIconModule } from 'ng-zorro-antd/icon';
-import { CvcOrgSubmitButtonDirective } from './org-submit-button.directive';
-import { CvcOrgSubmitButtonComponent } from './org-submit-button.type';
+import { CommonModule } from '@angular/common'
+import { NgModule } from '@angular/core'
+import { ReactiveFormsModule } from '@angular/forms'
+import { CvcOrganizationTagModule } from '@app/components/organizations/organization-tag/organization-tag.module'
+import { LetDirective, PushPipe } from '@ngrx/component'
+import { FormlyModule } from '@ngx-formly/core'
+import { ConfigOption } from '@ngx-formly/core/lib/models'
+import { NzAvatarModule } from 'ng-zorro-antd/avatar'
+import { NzButtonModule } from 'ng-zorro-antd/button'
+import { NzDropDownModule } from 'ng-zorro-antd/dropdown'
+import { NzIconModule } from 'ng-zorro-antd/icon'
+import { NzToolTipModule } from 'ng-zorro-antd/tooltip'
+import { CvcOrgSubmitButtonDirective } from './org-submit-button.directive'
+import { CvcOrgSubmitButtonComponent } from './org-submit-button.type'
 
 const typeConfig: ConfigOption = {
-  types: [{name: 'org-submit-button', component: CvcOrgSubmitButtonComponent}]
+  types: [
+    { name: 'org-submit-button', component: CvcOrgSubmitButtonComponent },
+  ],
 }
 
 @NgModule({
@@ -20,12 +24,15 @@ const typeConfig: ConfigOption = {
   imports: [
     CommonModule,
     ReactiveFormsModule,
-    LetDirective, PushPipe,
+    LetDirective,
+    PushPipe,
     FormlyModule.forChild(typeConfig),
     NzIconModule,
     NzAvatarModule,
     NzButtonModule,
+    NzToolTipModule,
     NzDropDownModule,
+    CvcOrganizationTagModule,
   ],
   exports: [CvcOrgSubmitButtonComponent, CvcOrgSubmitButtonDirective],
 })
